CMake fetch standard (#39)
Doing this PR as a test so we are on the same page. I'll do the same
work for BN and others once we settle down with this 'standard'.

- moved every `FetchContent_Declare` at the top of the file
- added `CONTENT_TO_FETCH` variable to hold content to really fetch
- a fetch for ios toolchain is separated. It doesn't look possible to
fetch the toolchain at the same place. I guess it's too late in CMake
pipeline
- set `TARGET` properties after fetch for obvious reason. This means
some tests are duplicated. IDK if it's possible to improve that (like
`if(NOT TARGET UrlLib AND JSRUNTIMEHOST_POLYFILL_XMLHTTPREQUEST)`)

---------

Co-authored-by: Gary Hsu <bghgary@users.noreply.github.com>