fix v8inspector crash at shutdown (#48)
`Stop ` method is called from 2 threads. the main when app stops or by
the thread itself. A mutex is added to guard concurrency. Then, when
Stop is called from the main thread, unique pointer is reset just after
but the thread itself is still using it. reset is removed from the main
thread. Thread is responsible for creating unique pointer and deleting
it.

Also removed `out_` because unused.