Restore lost xhr registration logic (#71)
`XmlHttpRequest` conditionally registers itself in the global scope, and
*used to* unconditionally register itself on the `_native` object used
for various things in `JsRuntime`. The latter part was lost when the
code was moved from the Babylon Native repo to the JsRuntimeHost repo
(see
https://github.com/BabylonJS/BabylonNative/pull/1260/files#diff-f5407a45027308c4ca372c3c5ad6ee33725fb2d39936521df64bf3c5b003b9cf).

As is, Babylon.js and Babylon React Native rely on this to ensure
Babylon Native's XHR is used rather than React Native's XHR (which is
very slow, very memory intensive, and does not support some features
like local file paths). See
https://github.com/BabylonJS/Babylon.js/blob/48cf7b374a7bbee9f3bef02f2992715ed683cf98/packages/dev/core/src/Misc/webRequest.ts#L12C1-L13C45.

Since JsRuntimeHost is supposed to be independent of Babylon related
code, having it work this way is probably not idea. However, this is
currently breaking shipping code, so we need to restore this
functionality for now and can consider an improved fix later.