Fixing various issues with build pipeline (#102)
Fixing a bunch of things that have broken in the build pipeline with
build agent updates over the last few months:
- MacOS & iOS: Use newer versions of Xcode (14 is no longer available)
- iOS: On either newer MacOS versions or Xcode versions, the emulator
has to be booted before the build, otherwise the build thinks there is
no available "destination" for the build.
- Android: SDKManager requires Java 8, which is no longer the default,
so it needs to be set explicitly before running SDKManager.
- Android: The emulator boot has errors if audio isn't disabled, so
disabling audio (and splash screen to make it faster).
- Android: The pipeline is setup to use an x86/x64 emulator, but the
newer MacOS build agent images are arm64. For now I just switched these
pipelines back to macos-13. Separately we can probably update the
pipeline to use arm64 emulator images.
- Android: The build would not deploy on the device. I tried locally and
got the same behavior. The only way I could get it to work was to
upgrade gradle. I did this with the upgrade assistant in Android Studio.
- Android: With the newer version of gradle, I also had to update to
Java 17 when running the gradle command.
- Android: Once the build was deploying, I was getting a runtime error
that seemed to be related to how the app interfaces with the emulator.
After some research, I found that I needed to upgrade the NDK to 23. I
assume this is because we are using a newer version of the emulator
itself (not the device image or API level).
- Android: With the gradle update, the gradle version is now in the test
result file names, but with spaces (e.g. `Pixel_API_27(AVD) -
8.1.0/aapt.1.ok.txt`), so I had to update the script that dumps the
contents of these files to the console.