Add PerfTrace API and add tracing to XmlHttpRequest (#125)
This PR depends on https://github.com/microsoft/arcana.cpp/pull/51

- Add a public PerfTrace API that aims to provide a platform agnostic
way of capturing perf intervals (just wraps arcana's trace_region). The
API can be used purely on the native side, but it also includes helper
functions for transferring ownership to JS, or taking ownership from JS.
This makes it easy to get perf intervals for operations that start in
native and end in JS, or vice versa.
- Also add a little more tracing to XmlHttpRequest.