Flesh out URL polyfill and add performance.now polyfill (#130)
- Flesh out the `URL` polyfill (moslty by Claude).
- Add a bunch more unit tests for `URL` (by Claude).
- Add a polyfill for `performance.now()` (by Claude).
- Add some unit tests for `performance.now()` (by Claude).
- Uncomment `JSGlobalContextSetInspectable` and add platform and os
version checks.
- Fix a bug where finalizers should be removed when `napi_remove_wrap`
is called (per https://nodejs.org/api/n-api.html#napi_remove_wrap).
Otherwise, a crash occurs if a constructor throws an error (which can
happen with URL when the passed in string is not a valid url). This also
addresses #82.
- Fix a bug where we try to assign to the result in `napi_remove_wrap`
when the `void** result` is null. We just need to check `result` before
assigning to it. `result` can be null in certain situations, such as if
a constructor throws (same scenario as above).
- Fix a small issue where the webpack watch command does not work (at
least on MacOS).